pyproject.toml
[project]
name = "dio-website-cms"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.8.0",
    "django-extensions>=4.1",
    "django-stubs[compatible-mypy]>=5.2.3",
    "wagtail>=7.1.1",
    "wagtailmenus>=4.0.4",
]

[tool.ruff]
line-length = 99
preview = true
exclude = [
    "migrations",
    "migration",
    "src/migrations",
    "alembic.ini",
    "**/__pycache__",
    "**/.mypy_cache",
    "**/.ruff_cache",
    "**/.venv"
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "EXE",
    "ISC",
    "ICN",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "INT",
    "ARG",
    "TD",
    "FIX",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "NPY",
    "AIR",
    "PERF",
    "FURB",
    "LOG",
    "RUF",
]
ignore = [
    "S105",
    "S101",
    "B008",
    "TRY201",
    "PLW3201",
    "N805",
    "PTH",
    "COM812",
    "ISC001",
    "TRY003",
    "PT013",
    "TID252",
    "RUF001",
    "RUF002",
    "RUF003",
    "ASYNC109",
]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "typings",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"typings" = ['typing']

[tool.ruff.lint.pylint]
max-args = 15
max-returns = 10
max-branches = 30

# -- MyPy --
[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
exclude = [
    "src/migration",
    "src/migrations",
    "__samples__",
    ".venv"
]

[tool.django-stubs]
django_settings_module = "myproject.settings"

# -- Vulture --
[tool.vulture]
exclude = [
    ".vscode/",
    "__samples__/",
    "tests",
    ".venv"
]
ignore_names = [
    "post",
    "put",
    "delete",
    "head"
]
